# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie-TV Show Finder API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/search": {
            "get": {
                "tags": [
                    "HomeController API routes"
                ],
                "summary": "Search",
                "description": "Search for movies and tv-shows or indiviadually",
                "operationId": "App\\Http\\Controllers\\HomeController::search",
                "parameters": [
                    {
                        "name": "searchTerm",
                        "in": "query",
                        "description": "Search term",
                        "required": true,
                        "example": "Friends"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data retrieved from an external API",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "movie-show": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MovieShow"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/details/{id}": {
            "get": {
                "tags": [
                    "HomeController API routes"
                ],
                "summary": "Movie-show details",
                "description": "Details of a particular movie or tv-show",
                "operationId": "App\\Http\\Controllers\\HomeController::findDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of movie/show",
                        "required": true,
                        "example": "176890"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item retrieved from an external API"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "MovieShow": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Title of a movie or show",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of a movie or show",
                        "type": "string"
                    },
                    "release_date": {
                        "description": "Release date of a movie or show",
                        "type": "string"
                    },
                    "vote_average": {
                        "description": "Average vote of a movie or show",
                        "type": "number",
                        "format": "float"
                    },
                    "poster_path": {
                        "description": "Path of a poster image",
                        "type": "string"
                    },
                    "popularity": {
                        "description": "Popularity of a particular movie or show",
                        "type": "number",
                        "format": "float"
                    },
                    "language": {
                        "description": "Spoken language of a movie or show",
                        "type": "string"
                    },
                    "original_title": {
                        "description": "Title in native language",
                        "type": "string"
                    },
                    "type": {
                        "description": "Either movie or tv-show",
                        "type": "string",
                        "enum": "[movie,tv-show]"
                    },
                    "api_Id": {
                        "description": "Id of a movie or show from external API",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MovieShow"
                }
            }
        }
    },
    "security": [
        []
    ]
}